核心读写状态原子化重构，新增CH32/FreeRTOS支持，支持可选Eigen依赖，多处风格与警告优化

+++ system/FreeRTOS/libxr_system.cpp
@@ -3,7 +3,8 @@
 
 static_assert(configTICK_RATE_HZ == 1000, "configTICK_RATE_HZ must be 1000");
 
-extern "C"  __attribute__((weak)) void vApplicationStackOverflowHook(TaskHandle_t xTask, char *pcTaskName)
+extern "C" __attribute__((weak)) void vApplicationStackOverflowHook(TaskHandle_t xTask,
+                                                                    char *pcTaskName)
 {
   static volatile const char *task_name = pcTaskName;
   UNUSED(task_name);
@@ -22,7 +23,8 @@ void LibXR::PlatformInit(uint32_t timer_pri, uint32_t timer_stack_depth)
   LibXR::Timer::priority_ = static_cast<LibXR::Thread::Priority>(timer_pri);
   LibXR::Timer::stack_depth_ = timer_stack_depth;
 
-  uint32_t time_need_to_catch_up = Timebase::GetMilliseconds() - xTaskGetTickCount();
+  int64_t time_need_to_catch_up = static_cast<int64_t>(Timebase::GetMilliseconds()) -
+                                  static_cast<int64_t>(xTaskGetTickCount());
 
   if (time_need_to_catch_up > 0)
   {
